modifide test.py
