+++ tests.py
@@ -1,24 +1,57 @@
-from main import BooksCollector
+class BooksCollector:
 
-# класс TestBooksCollector объединяет набор тестов, которыми мы покрываем наше приложение BooksCollector
-# обязательно указывать префикс Test
-class TestBooksCollector:
+    def __init__(self):
+        self.books_genre = {}
+        self.favorites = []
+        self.genre = ['Фантастика', 'Ужасы', 'Детективы', 'Мультфильмы', 'Комедии']
+        self.genre_age_rating = ['Ужасы', 'Детективы']
 
-    # пример теста:
-    # обязательно указывать префикс test_
-    # дальше идет название метода, который тестируем add_new_book_
-    # затем, что тестируем add_two_books - добавление двух книг
-    def test_add_new_book_add_two_books(self):
-        # создаем экземпляр (объект) класса BooksCollector
-        collector = BooksCollector()
 
-        # добавляем две книги
-        collector.add_new_book('Гордость и предубеждение и зомби')
-        collector.add_new_book('Что делать, если ваш кот хочет вас убить')
+    def add_new_book(self, name):
+        if not self.books_genre.get(name) and 0 < len(name) < 41:
+            self.books_genre[name] = ''
 
-        # проверяем, что добавилось именно две
-        # словарь books_rating, который нам возвращает метод get_books_rating, имеет длину 2
-        assert len(collector.get_books_rating()) == 2
 
-    # напиши свои тесты ниже
-    # чтобы тесты были независимыми в каждом из них создавай отдельный экземпляр класса BooksCollector()
+    def set_book_genre(self, name, genre):
+        if name in self.books_genre and genre in self.genre:
+            self.books_genre[name] = genre
+
+
+    def get_book_genre(self, name):
+        return self.books_genre.get(name)
+
+
+    def get_books_with_specific_genre(self, genre):
+        books_with_specific_genre = []
+        if self.books_genre and genre in self.genre:
+            for name, book_genre in self.books_genre.items():
+                if book_genre == genre:
+                    books_with_specific_genre.append(name)
+        return books_with_specific_genre
+
+
+    def get_books_genre(self):
+        return self.books_genre
+
+
+    def get_books_for_children(self):
+        books_for_children = []
+        for name, genre in self.books_genre.items():
+            if genre not in self.genre_age_rating and genre in self.genre:
+                books_for_children.append(name)
+        return books_for_children
+
+
+    def add_book_in_favorites(self, name):
+        if name in self.books_genre:
+            if name not in self.favorites:
+                self.favorites.append(name)
+
+
+    def delete_book_from_favorites(self, name):
+        if name in self.favorites:
+            self.favorites.remove(name)
+
+
+    def get_list_of_favorites_books(self):
+        return self.favorites
